This PR prefetches Parquet page headers so that they are read using one `FileSystem::Read` call, instead of having Thrift call into `FileSystem::Read` while decoding the page header, causing it to do many tiny reads (sometimes a few bytes at a time). This optimization will only happen when no other prefetches have happened (so it only applies to local reads), and the prefetch will get up to 256 bytes (which should cover most Parquet page headers).

When reading TPC-H SF10 lineitem using this query:
```sql
copy lineitem to 'lineitem_v1.parquet' (parquet_version v1); -- with V1 encodings
copy lineitem to 'lineitem_v2.parquet' (parquet_version v2); -- with V2 encodings
select any_value(columns(*)) from 'lineitem_v1.parquet';
select any_value(columns(*)) from 'lineitem_v2.parquet';
```
Performance improves like so:
V1: 0.48s -> 0.43s
V2: 0.36s -> 0.29s